deps: the cli tier's crate majors, and @types/node without TypeScript 7
Takes what #3 and #5 proposed, minus the one piece that is a migration rather than a bump.

cli/rs — getrandom 0.3->0.4, minreq 2->3, argon2 0.5->0.6, chacha20poly1305 0.10->0.11.
Two code changes the majors require:

  * chacha20poly1305 0.11 deprecates Array::from_slice. The key is a [u8; 32] so it takes
    From; the nonce is a &[u8] so it takes TryFrom, and the length error is now surfaced
    rather than panicking inside the codec. Both seal and open change.
  * minreq 3 types status_code as u16, so every `as u16` became a u16->u16 cast. All four
    are removed. Clippy flagged one; the compiler confirmed the other three by accepting
    their removal, which is the check that mattered — a cast clippy happened not to reach
    is still a cast that does nothing.

@types/node ^22 -> ^26 in all four TypeScript packages. TypeScript itself stays at ^5.7.

⛔ TypeScript 7.0.2 is deliberately NOT taken. #5 bundled it with @types/node and broke type
resolution across the whole tier — `Cannot find name 'process'`, `Cannot find name 'node:fs'`
— including in conformance/cli.ts, which is the harness the tri-core agreement runs through.
Measured before splitting: @types/node 26 compiles cleanly against TypeScript 5.9, so the
breakage is entirely the compiler jump. TS 7 is the native rewrite and deserves its own change
with its own verification, not a ride-along on a types bump.

Verified locally the way CI checks it: cargo fmt --all --check clean, cargo clippy
--all-targets --all-features -- -D warnings clean, and all 22 cli/rs tests pass.

+++ cli/rs/src/cmd/login.rs
@@ -546,7 +546,7 @@ fn fetch_login_request(audience: &str, id: &str) -> Result<LoginRequest, String>
         .map_err(|e| format!("login: could not reach {endpoint}: {e}"))?;
     let body = read_body(&response, &endpoint)?;
     if response.status_code != 200 {
-        return Err(login_http_error(response.status_code as u16, &body));
+        return Err(login_http_error(response.status_code, &body));
     }
     serde_json::from_str(&body)
         .map_err(|e| format!("login: the service's request is not the expected JSON: {e}"))
@@ -577,7 +577,7 @@ fn post_answer(audience: &str, id: &str, answer: &LoginAnswer) -> Result<(u16, S
         .send()
         .map_err(|e| format!("login: could not reach {endpoint}: {e}"))?;
     let body = response.as_str().unwrap_or("").to_string();
-    Ok((response.status_code as u16, body))
+    Ok((response.status_code, body))
 }
 
 /// The response body as text. A body that is not UTF-8 is a service speaking something this
@@ -892,7 +892,7 @@ fn post_offer(audience: &str, offer: &OfferBody) -> Result<OfferResponse, String
         .map_err(|e| format!("login: could not reach {endpoint}: {e}"))?;
     let body = read_body(&response, &endpoint)?;
     if response.status_code != 201 {
-        return Err(login_http_error(response.status_code as u16, &body));
+        return Err(login_http_error(response.status_code, &body));
     }
     serde_json::from_str(&body)
         .map_err(|e| format!("login: the service's offer is not the expected JSON: {e}"))
@@ -979,7 +979,7 @@ fn poll_offer(
             429 => continue,
             404 => return Err("login: the offer expired before the page took it".to_string()),
             200 => {}
-            other => return Err(login_http_error(other as u16, &body)),
+            other => return Err(login_http_error(other, &body)),
         }
         let read: OfferRead = serde_json::from_str(&body)
             .map_err(|e| format!("login: the service's offer is not the expected JSON: {e}"))?;
